requirements.txt
beautifulsoup4 == 4.13.4
cloudscraper == 1.2.71
python-telegram-bot == 22.0